Consistently round table row heights.
Do in LogicalHeightForRow() as in CalcRowLogicalHeight() - i.e. round
instead of floor, to convert from LayoutUnit to int. We calculated the
pagination strut incorrectly for the line that followed, because it was
based on the floored LayoutUnit height of the preceding table row.
Then, later on, it got adjusted (rounded from LayoutUnit), and the
balancing code got confused, because the line had a pagination strut,
but didn't appear to be exactly at the start of a column (there should
be no strut in such a case).

Bug: 908247
Change-Id: I11fa5430b9ff6e83b20dc25ce1d945e54b80f6ea